requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
flask==2.3.3
supabase>=2.0.0
pgvector>=0.2.0